requirements.txt
numpy
pandas
contractions
nltk
gensim
wordcloud
seaborn
matplotlib
joblib
tensorflow
keras
djangorestframework
django
django-allauth
fitz
